Rename file ("sso" -> "fairdata_sso")
